Refactored Asset Component
